+++ contracts/fx-portal/tunnel/ERC20/FxERC20FixedSupplyChildTunnel.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.28;
 
-import {IFxERC20} from "./../../token/ERC20/interfaces/IFxERC20.sol";
 import {IFxERC20FixedSupply} from "./../../token/ERC20/interfaces/IFxERC20FixedSupply.sol";
 import {IForwarderRegistry} from "@animoca/ethereum-contracts/contracts/metatx/interfaces/IForwarderRegistry.sol";
 import {IERC20, SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";